Changed index of Nav to export from Nav.js
